moving review request and response types into their own file
