+++ src/services/products/reviews/types/reviewRequestTypes.ts
@@ -1,0 +1,15 @@
+export interface CreateReviewRequest {
+  review: {
+    rating: number,
+    max_rating: number,
+    review_text: string
+  }
+}
+
+export interface UpdateReviewRequest {
+  review: {
+    rating?: number,
+    max_rating?: number,
+    review_text?: string
+  }
+}
