+++ src/services/products/reviews/types/reviewResponseTypes.ts
@@ -1,18 +1,3 @@
-export interface CreateReviewRequest {
-  review: {
-    rating: number,
-    max_rating: number,
-    review_text: string
-  }
-}
-
-export interface UpdateReviewRequest {
-  review: {
-    rating?: number,
-    max_rating?: number,
-    review_text?: string
-  }
-}
 export interface GetReviewResponse {
   uuid: string,
   rating_percentage: number,
